requirements.txt
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.8.2
